requirements.txt
discord.py[voice]==2.4.0
aiohttp==3.9.5
aiosqlite==0.20.0
